+++ src/components/Button/components/LogOutBtn/LogOutBtn.module.scss
@@ -1,0 +1,10 @@
+@import './../../Button.module.scss';
+
+.logOutBtn{
+  border: solid 1px $logout-color;
+  color: $logout-color;
+  &:hover{
+    color:$hover-color;
+    background-color: $logout-color;
+  }
+}
