+++ src/components/Button/components/LogOutBtn/index.js
@@ -1,0 +1,1 @@
+export {default } from './LogOutBtn';
