+++ src/components/Main/components/Content/components/HalfWidthLayout/HalfWidthLayout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "./HalfWidthLayout.module.scss";
-import Button from "../Button";
+import Button from "./../../../../../Button";
 import StudentAssignment from './../StudentAssignment';
 import Profile from './../Profile';
 import { connect } from "react-redux";
@@ -35,6 +35,7 @@ const HalfWidthLayout = (props) => {
               <Button type={"DELETE"} />
               <Button type={"UPDATE"} />
               <Button type={"CREATE"} />
+              <Button type={"LOGOUT"} />
             </React.Fragment>
           );
         case "Context":
