+++ src/components/Nav/components/NavSideBar/components/Canvas/components/Assignment/Assignment.scss
@@ -1,0 +1,36 @@
+@import './../../../../NavSideBar-variable';
+
+.wrapper{
+  display: flex;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.active{
+  color: $dark-color;
+   background-color: $light-dark-color;
+}
+.inActive{
+  color: $light-dark-color;
+}
+
+.icon{
+  width:$icon-width;
+  height:40px;
+  font-size: 1.2em;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+ 
+}
+.title{
+  height:40px;
+  display: flex;
+  align-items: center;
+  h3{
+  color: $white-color;
+  }
+}
+.hide{
+  display: none;
+}
