+++ src/components/Nav/components/NavSideBar/components/Canvas/components/Assignment/index.js
@@ -1,0 +1,1 @@
+export {default} from "./CourseMarket";
