+++ src/components/Popup/Create.jsx
@@ -1,7 +1,6 @@
 import React from "react";
-import styles from './form.module.scss';
-import axios from 'axios';
-
+import styles from "./form.module.scss";
+import studentService from "./../../apis/studentService";
 
 // Post object would need a name
 
@@ -12,61 +11,50 @@ class Create extends React.Component {
       email: "",
       firstName: "",
       lastName: "",
-      respDto:{},
+      submitted: false,
     };
     this.onEmailTyping = this.onEmailTyping.bind(this);
     this.onFirstNameTyping = this.onFirstNameTyping.bind(this);
     this.onLastNameTyping = this.onLastNameTyping.bind(this);
-    this.onSubmit = this.onSubmit.bind(this);
-  }
-  
-  //目前是，如果拿不回来你们的窗口不会关。所以要强行关掉
-  async sendDto(){
-    const newStudent = {
-      email:this.state.email,
-      firstName:this.state.firstName,
-      lastName:this.state.lastName,
-    }
-    const studentDTO = JSON.stringify(newStudent);
-
-
-    //这里的格式千万要注意。不要放一个{ }到post 的第二个argument。因为这样后端的格式就不是JSON
-      console.log(studentDTO);
-     await axios
-       .post(
-         "http://localhost:8080/users/students",
-         studentDTO ,
-         {
-           headers: {
-             "Content-Type": "application/json",
-           },
-         }
-       )
-       .then((res) => this.setState({ respDto: res.data }));
-    console.log(this.state.respDto);
-  }
-  async onSubmit(event){
-    event.preventDefault();
-    await this.sendDto();
-    this.props.onClick(event);
-
-    // this is to resemble sending JSON to the server
-
+    this.save = this.save.bind(this);
   }
 
-  onEmailTyping(event){
+  save() {
+    var data = {
+      email: this.state.email,
+      firstName: this.state.firstName,
+      lastName: this.state.lastName,
+    };
+
+    studentService
+      .create(data)
+      .then((response) => {
+        this.setState({
+          email: response.data.email,
+          firstName: response.data.firstName,
+          lastName: response.data.lastName,
+          submitted: true,
+        });
+        console.log(response.data);
+      })
+      .catch((e) => {
+        console.log(e);
+      });
+  }
+
+  onEmailTyping(event) {
     console.log(event.target.email);
     this.setState({
       email: event.target.value,
     });
   }
-  onFirstNameTyping(event){
+  onFirstNameTyping(event) {
     console.log(event.target.firstName);
     this.setState({
       firstName: event.target.value,
     });
   }
-  onLastNameTyping(event){
+  onLastNameTyping(event) {
     console.log(event.target.lastName);
     this.setState({
       lastName: event.target.value,
@@ -75,7 +63,11 @@ class Create extends React.Component {
 
   render() {
     return (
-      <form className={styles.form} onSubmit={this.onSubmit}>
+      <div>
+        {this.state.submitted ?(<div>
+        <h3 className={styles.title}>Student has been created Successfully!</h3>
+        </div>):(
+      <form className={styles.form}>
         <h3 className={styles.title}>Create</h3>
         <div className={styles.control}>
           <label>Email:</label>
@@ -103,14 +95,16 @@ class Create extends React.Component {
             required
           />
         </div>
-
         <button
+          onClick={this.save}
           className={styles.button}
           type="submit"
         >
           Submit
         </button>
       </form>
+      )}
+      </div>
     );
   }
 }
