+++ src/store/authentication/actions/LogOut.js
@@ -1,0 +1,5 @@
+import { LOGOUT } from "../../type";
+
+export default () => ({
+  type: LOGOUT,
+});
