--- src/store/authentication/type.js
@@ -1,14 +1,0 @@
-
-//role reducer
-export const CHANGE_ROLE = "CHANGE_ROLE";
-export const SET_LOGIN_STATUS = "SET_LOGIN_STATUS";
-export const SET_CURRENT_DIRECTOR = "SET_CURRENT_DIRECTOR";
-
-//link
-export const ADD_HEADER_HISTORY = "ADD_HEADER_HISTORY";
-export const CLICK_HEADER = "CLICK_HEADER";
-
-//canvas
-export const SET_CANVAS_CURRENT = "SET_CANVAS_CURRENT";
-export const TOGGLE_CANVAS = "TOGGLE_CANVAS";
-export const SET_CANVAS_STATUS = "SET_CANVAS_STATUS";
