+++ src/store/type.js
@@ -3,6 +3,7 @@
 export const CHANGE_ROLE = "CHANGE_ROLE";
 export const SET_LOGIN_STATUS = "SET_LOGIN_STATUS";
 export const SET_CURRENT_DIRECTOR = "SET_CURRENT_DIRECTOR";
+export const LOGOUT = "LOGOUT";
 
 //link
 export const ADD_HEADER_HISTORY = "ADD_HEADER_HISTORY";
